+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Fix touch event targeting in nested windows.
 * Fix `MOUSE.position` not tracking nested windows.
 * Fix context menu not opening in nested windows.
 * Add `AnyEventArgs::backtrace` for tracking event args sources.